Created minimally-functional version of game loop
This is a very simple game loop that will remain open safely until the player closes the game window, whether by hitting Q or closing the terminal window.

In addition, the README has been updated with information on how to get state and input information from BearLibTerminal, as well as a list of events and keys that BLT can observe. There is much more comprehensive information available at
   http://foo.wyrd.name/en:bearlibterminal:reference:input

 On branch gameLoopPrototype
 Changes to be committed:
	modified:   README
	modified:   src/engine.cpp
	modified:   src/gui.cpp